Most annoying bug ever.
Importing one of our modules (other than os and sys, which I double checked) causes a file to be opened.
When splatd closes all its file descriptors, a reference to that file descriptor remains, referenced somewhere in the bowels of the interpreter. Later on, when the LDAP connection is lost, select() is called on that file descriptor, which just so happens to refer to the now-dead LDAP connection. select() immediately returns, and then a read() is called on the FD, which returns 0. This continues in a tight loop until the LDAP connection is re-initialized.

Now, instead, we close all our file descriptors *immediately*, before importing anything other than os and sys.
Python really needs to provide an intelligent way to close all file descriptors for a daemon.

Python--.