Move _parseBooleanOption() from purgeUser plugin to its superclass, and have other plugins use that. Also used string object lower() method rather than deprecated string.lower(str) class method.
In purgeUser, Don't require accountStatus, as it is specific to qmailAccount
LDAP object classes. Removed os.umask in _archiveHomeDir() because we'll create
the archive file with the right permissions a different way. Finally, some style
fixups like using mixedCase for WriterContext's variables, and putting imports
each on their own line.